Housekeeping.
-> Corrected typos involving lithuanian in a number of places in LanguageModelConversions.cs. Corrected typos where portuguese was typed out as portugueuse and portugal as portgual in Language.cs, LanguageModelConversions.cs. and EmbedGenerator.cs

-> Corrected numerous different typos.
In EmbedGenerator.cs,
the variable unkownLanguageEmbedBuilder in the GenerateUnkownLanguageEmbedBuilder method was renamed to unknownLanguageEmbedBuilder. In the GeneratePossibleLanguages method, the same previously covered mistakes in paragraph 1 were fixed. In the documentation for the class, the word "embeded" was a mistake and was corrected to be embedded. In the GenerateReconnectionEmbed function, the parameter had a typo and was called "wasSuccesful". It has been fixed and is now "wasSuccessful".
In TranslationService.cs in the documentation for the method TranslatorSetUp, there was a typo on the word 'unsuccessful'. It has been fixed.
In Utils.cs: in the documentation for the method DivideUpTextIntoFragmentsNicely, three typos were fixed. The word 'specified' in the summary is now correct. In the documentation for the maxLength parameter, the typo on the word 'value' was fixed. Lastly, the word beginning now has two ns instead of one in the documentation for the parameter initialText.

-> Object initializations for embeds in EmbedGenerator.cs are now done via object initializers and not with separate constructor and attribute value assignments.

-> Unused imports have been removed from the Program.cs, LanguageModelConversions.cs, TranslationModule.cs and TranslationService.cs files

-> namespaces have been adjusted in many places as a temporary measure for now. The affected files are Utils.cs, Program.cs, LanguageModelConversions.cs, Language.cs and EmbedGenerator.cs

-> The Program class in Program.cs is now a static class.

-> The translator object in TranslationService.cs having been changed to being private, it is now called _translator and not Translator to fit with this change.

-> The README.md has had a serious makeover and is now much more functional and useful.