Changed the namespace for the TranslationModule.cs from Template.Modules to Templates.Modules.Translation as the module files have been moved to the respective module directories.
Added documentation for the LanguageModelConversions static class in the LanguageModelConversions.cs file.

The color for generated embeds in EmbedGenerator.cs for the TranslationModule is now dynamic: red if the color is unknown, dark green for a successful translation. The color is orange if the translation limit has been reached. The color is Red if the text to translate is empty. It is DarkRed if the api connection error embed is generated. Finally for the reconnection embed, the color is dark green if the connection was successful and dark red if the connection was not successful.