Accidentally omitted of the changes to EmbedGenerator.cs in the commit.
The color for generated embeds in EmbedGenerator.cs for the TranslationModule is now dynamic: red if the color is unknown, dark green for a successful translation. The color is orange if the translation limit has been reached. The color is Red if the text to translate is empty. It is DarkRed if the api connection error embed is generated. Finally for the reconnection embed, the color is dark green if the connection was successful and dark red if the connection was not successful.